backend/requirements.txt
psycopg2 == 2.8.5
Django == 3.1.6
djangorestframework == 3.12.0
coreapi == 2.3.3
drf-spectacular == 0.13.2
requests == 2.23.0
Pillow == 8.4
cryptography == 2.1.4
django-filter == 21.1
factory-boy == 2.12.0
coverage == 5.1.0
django-debug-toolbar == 2.2

